Move rewriteProjectionAttributeAccesses
