+++ arangod/Aql/Optimizer/Rule/MaterializeForEnumerateNear.cpp
@@ -28,7 +28,7 @@
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionNode/MaterializeRocksDBNode.h"
 #include "Aql/Optimizer.h"
-#include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Utils/RewriteProjectionAttributeAccesses.h"
 #include "Aql/Optimizer/Utils/FindProjections.h"
 #include "Aql/Projections.h"
 #include "Containers/FlatHashSet.h"
@@ -95,7 +95,7 @@ void materializeForEnumerateNear(Optimizer* opt,
 
     if (indexCoversProjections || filterLoadsDocument) {
       if (!projections.empty()) {
-        utils::rewriteProjectionAttributeAccesses(
+        optimizer::rewriteProjectionAttributeAccesses(
             *plan, vectorNode, vectorNode->outVariable(), projections,
             /*index*/ 0);
         vectorNode->setProjections(std::move(projections));
