+++ arangod/Aql/Optimizer/Rule/OptimizeProjections.cpp
@@ -30,16 +30,12 @@
 #include "Aql/ExecutionNode/MaterializeRocksDBNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Optimizer.h"
-#include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Utils/RewriteProjectionAttributeAccesses.h"
 #include "Aql/Optimizer/Utils/FindProjections.h"
 #include "Aql/Projections.h"
 #include "Aql/Variable.h"
 #include "Containers/FlatHashSet.h"
 #include "Containers/SmallVector.h"
-
-#include <span>
-#include <string_view>
-#include <vector>
 
 namespace arangodb::aql {
 using EN = ExecutionNode;
@@ -56,8 +52,8 @@ void optimizeProjections(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
     if (p.empty()) {
       return false;
     }
-    utils::rewriteProjectionAttributeAccesses(*plan, self, searchVariable, p,
-                                              index);
+    optimizer::rewriteProjectionAttributeAccesses(*plan, self, searchVariable,
+                                                  p, index);
     return true;
   };
